[BACKEND][ESTRUTURA DO PROJETO] Separado metodos relacionado a lógico de negócio em classes Service
