+++ api/src/main/java/itau/canais/api/modules/dto/DadosConta.java
@@ -1,4 +1,4 @@
-package itau.canais.api.dto;
+package itau.canais.api.modules.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import org.hibernate.validator.constraints.br.CPF;
