+++ api/src/main/java/itau/canais/api/modules/dto/DadosListagemClientes.java
@@ -1,0 +1,12 @@
+package itau.canais.api.modules.dto;
+
+import itau.canais.api.modules.entities.Cliente;
+
+public record DadosListagemClientes(
+        String nome,
+        String cpf){
+
+    public DadosListagemClientes(Cliente cliente){
+        this(cliente.getNome(), cliente.getCpf());
+    }
+}
