+++ api/src/main/java/itau/canais/api/modules/exceptions/ClienteNotFoundException.java
@@ -1,4 +1,4 @@
-package itau.canais.api.exceptions;
+package itau.canais.api.modules.exceptions;
 
 public class ClienteNotFoundException extends Throwable {
 }
