+++ api/src/main/java/itau/canais/api/modules/exceptions/SaldoInsuficienteException.java
@@ -1,4 +1,4 @@
-package itau.canais.api.controller;
+package itau.canais.api.modules.exceptions;
 
 public class SaldoInsuficienteException extends Throwable {
 }
