+++ api/src/main/java/itau/canais/api/modules/repositories/ContaRepository.java
@@ -1,16 +1,13 @@
-package itau.canais.api.repositories;
+package itau.canais.api.modules.repositories;
 
-import itau.canais.api.entities.Cliente;
-import itau.canais.api.entities.Conta;
-import jakarta.validation.constraints.NotNull;
+import itau.canais.api.modules.dto.DadosConta;
+import itau.canais.api.modules.entities.Conta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.math.BigDecimal;
 
 @Repository
 public interface ContaRepository extends JpaRepository<Conta, String> {
@@ -18,8 +15,6 @@ public interface ContaRepository extends JpaRepository<Conta, String> {
     @Query("SELECT e FROM Conta e WHERE e.agencia = :agencia AND e.nconta = :nconta")
     public Conta buscarContaByAgenciaConta(@Param("agencia") String agencia, @Param("nconta") String nconta);
 
-//    @Query("SELECT e FROM Conta e WHERE e.agencia = :agencia AND e.nconta = :nconta")
-//    Optional<Conta> buscarContaByAgenciaConta(@Param("agencia") String agencia, @Param("nconta") String nconta);
-
+    void transferirSaldo(DadosConta contaOrigem, DadosConta contaDestino, BigDecimal valor);
 }
